feat(suite): move MCP server from debug to experimental features
- Register 'mcp-server' as experimental feature with toggle, docs link,
  and onToggle callback to start/stop the server
- Move McpServer UI from SettingsDebug to SettingsGeneral with per-client
  config snippets: Claude Code (CLI), Claude Desktop (mcp-remote bridge),
  and other MCP clients (JSON)
- Add stablecoin contract references and ERC-20 transfer instructions
  to MCP skill
- Update all docs to reference Experimental Features instead of Debug
- Add mcp-remote troubleshooting for Claude Code SDK OAuth regression
